+++ src/app/models/users/profile.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Enum, ForeignKey, String, func
+from sqlalchemy import DateTime, Enum, ForeignKey, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database.base import Base
@@ -12,7 +12,7 @@ class Profile(Base):
     __tablename__ = "profiles"
 
     account_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         ForeignKey("accounts.id", name="fk_profiles_account"),
         primary_key=True,
     )
